Fix server startup flow
